jobLogging/requirements.txt
pycups
datetime
ConfigParser
tabulate
